Masonry reverts (#339)
* Revert "Masonry: update rendering states to match with items props (#319)"

This reverts commit 1001767da842f0bb5150916388be90babeb52721.

* Revert "Masonry: add missing defaultProps and handleResize (#313)"

This reverts commit 865e67fe30debbcb846c704b2bc3b1b57a2edde0.

* Revert "Masonry: MasonryInfinite for infinite fetching (#307)"

This reverts commit f6c683b66b2d8b0ec87db283418459e87160a21f.
