+++ packages/gestalt/src/Masonry.js
@@ -2,6 +2,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import debounce from './debounce.js';
+import FetchItems from './FetchItems.js';
 import styles from './Masonry.css';
 import ScrollContainer from './ScrollContainer.js';
 import throttle from './throttle.js';
@@ -16,7 +17,7 @@ import {
   DefaultLayoutSymbol,
   UniformRowLayoutSymbol,
 } from './legacyLayoutSymbols.js';
-import defaultLayout, { type Position } from './defaultLayout.js';
+import defaultLayout from './defaultLayout.js';
 import uniformRowLayout from './uniformRowLayout.js';
 import fullWidthLayout from './fullWidthLayout.js';
 import LegacyMasonryLayout from './layouts/MasonryLayout.js';
@@ -28,9 +29,7 @@ type Layout =
   | LegacyMasonryLayout
   | LegacyUniformRowLayout;
 
-export type MeasurementState = 'idle' | 'measuring';
-
-export type Props<T> = {|
+type Props<T> = {|
   columnWidth?: number,
   comp: React.ComponentType<{
     data: T,
@@ -42,12 +41,6 @@ export type Props<T> = {|
   items: Array<T>,
   measurementStore: Cache<T, *>,
   minCols: number,
-  // Content layer and Viewport layer is as defined in Collection.
-  onVirtualizationWindowUpdate?: (
-    content: Position,
-    viewport: Position
-  ) => void,
-  onAutoMeasuringUpdate?: (state: MeasurementState) => void,
   layout?: Layout,
   // Support legacy loadItems usage.
   // TODO: Simplify non falsey flowtype.
@@ -66,12 +59,8 @@ export type Props<T> = {|
 
 type State<T> = {|
   hasPendingMeasurements: boolean,
-  height: number,
+  isFetching: boolean,
   items: Array<T>,
-  itemsToMeasure: Array<T>,
-  itemsToRender: Array<T>,
-  measuringPositions: Array<Position>,
-  renderPositions: Array<Position>,
   scrollTop: number,
   width: ?number,
 |};
@@ -83,87 +72,11 @@ const VIRTUAL_BUFFER_FACTOR = 0.7;
 
 const layoutNumberToCssDimension = n => (n !== Infinity ? n : undefined);
 
-function layoutClass<T>(
-  {
-    columnWidth,
-    flexible,
-    gutterWidth: gutter,
-    layout,
-    measurementStore,
-    minCols,
-  }: Props<T>,
-  { width }: State<T>
-) {
-  if (flexible && width !== null) {
-    return fullWidthLayout({
-      gutter,
-      cache: measurementStore,
-      minCols,
-      idealColumnWidth: columnWidth,
-      width,
-    });
-  }
-  if (
-    layout === UniformRowLayoutSymbol ||
-    layout instanceof LegacyUniformRowLayout
-  ) {
-    return uniformRowLayout({
-      cache: measurementStore,
-      columnWidth,
-      gutter,
-      minCols,
-      width,
-    });
-  }
-  return defaultLayout({
-    cache: measurementStore,
-    columnWidth,
-    gutter,
-    minCols,
-    width,
-  });
-}
-
-function statesForRendering<T>(props: Props<T>, state: State<T>) {
-  const { measurementStore, minCols } = props;
-  const { items } = state;
-
-  // Full layout is possible
-  // $FlowFixMe https://github.com/facebook/flow/issues/6151
-  const itemsToRender = items.filter(
-    item => item && measurementStore.has(item)
-  );
-
-  const layout = layoutClass(props, state);
-  const renderPositions = layout(itemsToRender);
-  // Math.max() === -Infinity when there are no renderPositions
-  const height = renderPositions.length
-    ? Math.max(...renderPositions.map(pos => pos.top + pos.height))
-    : 0;
-
-  // $FlowFixMe https://github.com/facebook/flow/issues/6151
-  const itemsToMeasure = items
-    .filter(item => item && !measurementStore.has(item))
-    .slice(0, minCols);
-  const measuringPositions = layout(itemsToMeasure);
-
-  return {
-    height,
-    itemsToRender,
-    itemsToMeasure,
-    measuringPositions,
-    renderPositions,
-  };
-}
-
-/**
- * TODO this should be renamed to MasronyBeta or something else so it is clear
- * that this is not the exported Masonry.
- *
- * The goal is to eventually not have any scroll fetching concerns in this component.
- * The name is kept for now to have an easier time seeing the diffs.
- */
 export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
+  static createMeasurementStore() {
+    return new MeasurementStore();
+  }
+
   /**
    * Delays resize handling in case the scroll container is still being resized.
    */
@@ -186,8 +99,6 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
     this.setState({
       scrollTop: getScrollPos(scrollContainer),
     });
-
-    this.handleVirtualizationWindowUpdate();
   });
 
   measureContainerAsync = debounce(() => {
@@ -280,13 +191,9 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       hasPendingMeasurements: props.items.some(
         item => !!item && !props.measurementStore.has(item)
       ),
-      height: 0,
-      itemsToRender: [],
-      itemsToMeasure: [],
+      isFetching: false,
       // eslint-disable-next-line react/no-unused-state
       items: props.items,
-      measuringPositions: [],
-      renderPositions: [],
       scrollTop: 0,
       width: undefined,
     };
@@ -312,11 +219,6 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       scrollTop,
       width: this.gridWrapper ? this.gridWrapper.clientWidth : prevState.width,
     }));
-
-    // need to make sure parent component has the correct pending measurement value
-    this.handleOnAutoMeasuringUpdate(
-      this.state.hasPendingMeasurements ? 'measuring' : 'idle'
-    );
   }
 
   componentDidUpdate(prevProps: Props<T>, prevState: State<T>) {
@@ -331,30 +233,15 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
     const hasPendingMeasurements = items.some(
       item => !!item && !measurementStore.has(item)
     );
-
-    if (hasPendingMeasurements && !prevState.hasPendingMeasurements) {
-      this.handleOnAutoMeasuringUpdate('measuring');
-    } else if (!hasPendingMeasurements && prevState.hasPendingMeasurements) {
-      this.handleOnAutoMeasuringUpdate('idle');
-    }
-    this.handleVirtualizationWindowUpdate();
-
     if (
       hasPendingMeasurements ||
       hasPendingMeasurements !== this.state.hasPendingMeasurements ||
       prevState.width == null
     ) {
       this.insertAnimationFrame = requestAnimationFrame(() => {
-        const renderingStates = statesForRendering(this.props, this.state);
         this.setState({
           hasPendingMeasurements,
-          ...renderingStates,
         });
-      });
-    } else if (hasPendingMeasurements || prevState.items !== items) {
-      this.insertAnimationFrame = requestAnimationFrame(() => {
-        const renderingStates = statesForRendering(this.props, this.state);
-        this.setState({ ...renderingStates });
       });
     }
   }
@@ -380,14 +267,8 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
     // whenever we're receiving new props, determine whether any items need to be measured
     // TODO - we should treat items as immutable
     const hasPendingMeasurements = items.some(
-      item => item && !measurementStore.has(item)
+      item => !measurementStore.has(item)
     );
-
-    const newState: State<T> = {
-      ...state,
-      hasPendingMeasurements,
-      items,
-    };
 
     // Shallow compare all items, if any change reflow the grid.
     for (let i = 0; i < items.length; i += 1) {
@@ -397,7 +278,7 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
         return {
           hasPendingMeasurements,
           items,
-          ...statesForRendering(props, newState),
+          isFetching: false,
         };
       }
 
@@ -411,7 +292,7 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
         return {
           hasPendingMeasurements,
           items,
-          ...statesForRendering(props, newState),
+          isFetching: false,
         };
       }
     }
@@ -421,7 +302,7 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       return {
         hasPendingMeasurements,
         items,
-        ...statesForRendering(props, newState),
+        isFetching: false,
       };
     }
     if (hasPendingMeasurements !== state.hasPendingMeasurements) {
@@ -429,7 +310,6 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       return {
         hasPendingMeasurements,
         items,
-        ...statesForRendering(props, newState),
       };
     }
 
@@ -445,32 +325,15 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
     this.scrollContainer = ref;
   };
 
-  handleVirtualizationWindowUpdate = () => {
-    const { height, width } = this.state;
-    if (
-      typeof this.props.onVirtualizationWindowUpdate === 'function' &&
-      this.containerHeight
-    ) {
-      const viewport = {
-        top: this.state.scrollTop,
-        left: 0,
-        height: this.containerHeight,
-        width: width || 0,
-      };
-      const content = {
-        top: this.containerOffset,
-        left: 0,
-        height,
-        width: width || 0,
-      };
-
-      this.props.onVirtualizationWindowUpdate(content, viewport);
-    }
-  };
-
-  handleOnAutoMeasuringUpdate = (state: MeasurementState) => {
-    if (this.props.onAutoMeasuringUpdate) {
-      this.props.onAutoMeasuringUpdate(state);
+  fetchMore = () => {
+    const { loadItems } = this.props;
+    if (loadItems && typeof loadItems === 'function') {
+      this.setState(
+        {
+          isFetching: true,
+        },
+        () => loadItems({ from: this.props.items.length })
+      );
     }
   };
 
@@ -510,7 +373,6 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
   reflow() {
     this.props.measurementStore.reset();
     this.measureContainer();
-    this.handleVirtualizationWindowUpdate();
     this.forceUpdate();
   }
 
@@ -571,18 +433,43 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       columnWidth,
       comp: Component,
       flexible,
+      gutterWidth: gutter,
       measurementStore,
       items,
+      minCols,
     } = this.props;
-    const {
-      hasPendingMeasurements,
-      height,
-      itemsToMeasure,
-      itemsToRender,
-      measuringPositions,
-      renderPositions,
-      width,
-    } = this.state;
+    const { hasPendingMeasurements, width } = this.state;
+
+    let layout;
+    if (flexible && width !== null) {
+      layout = fullWidthLayout({
+        gutter,
+        cache: measurementStore,
+        minCols,
+        idealColumnWidth: columnWidth,
+        width,
+      });
+    } else if (
+      this.props.layout === UniformRowLayoutSymbol ||
+      this.props.layout instanceof LegacyUniformRowLayout
+    ) {
+      layout = uniformRowLayout({
+        cache: measurementStore,
+        columnWidth,
+        gutter,
+        minCols,
+        width,
+      });
+    } else {
+      layout = defaultLayout({
+        cache: measurementStore,
+        columnWidth,
+        gutter,
+        minCols,
+        width,
+      });
+    }
+
     let gridBody;
     if (width == null && hasPendingMeasurements) {
       // When hyrdating from a server render, we don't have the width of the grid
@@ -624,12 +511,25 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
       // div to collect the width for layout
       gridBody = <div style={{ width: '100%' }} ref={this.setGridWrapperRef} />;
     } else {
+      // Full layout is possible
+      const itemsToRender = items.filter(
+        item => item && measurementStore.has(item)
+      );
+      const itemsToMeasure = items
+        .filter(item => item && !measurementStore.has(item))
+        .slice(0, minCols);
+
+      const positions = layout(itemsToRender);
+      const measuringPositions = layout(itemsToMeasure);
+      // Math.max() === -Infinity when there are no positions
+      const height = positions.length
+        ? Math.max(...positions.map(pos => pos.top + pos.height))
+        : 0;
       gridBody = (
         <div style={{ width: '100%' }} ref={this.setGridWrapperRef}>
           <div className={styles.Masonry} style={{ height, width }}>
             {itemsToRender.map((item, i) =>
-              // $FlowFixMe this is the right definition, it an Array<T>
-              this.renderMasonryComponent(item, i, renderPositions[i])
+              this.renderMasonryComponent(item, i, positions[i])
             )}
           </div>
           <div className={styles.Masonry} style={{ width }}>
@@ -665,6 +565,18 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
               );
             })}
           </div>
+
+          {this.scrollContainer && (
+            <FetchItems
+              containerHeight={this.containerHeight}
+              fetchMore={this.fetchMore}
+              isFetching={
+                this.state.isFetching || this.state.hasPendingMeasurements
+              }
+              scrollHeight={height}
+              scrollTop={this.state.scrollTop}
+            />
+          )}
         </div>
       );
     }
