+++ packages/gestalt/src/defaultLayout.js
@@ -1,12 +1,7 @@
 // @flow
 import type { Cache } from './Cache.js';
 
-export type Position = {
-  top: number,
-  left: number,
-  width: number,
-  height: number,
-};
+type Position = { top: number, left: number, width: number, height: number };
 
 const mindex = arr => {
   let idx = 0;
