+++ packages/gestalt/src/index.js
@@ -20,7 +20,6 @@ import Layer from './Layer.js';
 import Letterbox from './Letterbox.js';
 import Link from './Link.js';
 import Mask from './Mask.js';
-import MasonryInfinite from './MasonryInfinite.js';
 import Masonry from './Masonry.js';
 import MasonryDefaultLayout from './layouts/MasonryLayout.js';
 import MasonryUniformRowLayout from './layouts/UniformRowLayout.js';
@@ -66,8 +65,7 @@ export {
   Letterbox,
   Link,
   Mask,
-  MasonryInfinite as Masonry,
-  Masonry as MasonryBeta,
+  Masonry,
   MasonryDefaultLayout,
   MasonryUniformRowLayout,
   Modal,
