+++ test/containers/MasonryExample.js
@@ -24,8 +24,6 @@ const getRandomNumberGenerator = seed => {
 };
 
 export default class MasonryExample extends React.Component {
-  autoMeasurementState = 'measuring'; // MeasurementState
-
   randomNumberSeed = 0;
 
   constructor(props) {
@@ -59,7 +57,7 @@ export default class MasonryExample extends React.Component {
       window.RESIZE_MEASUREMENT_DONE = false;
       setTimeout(() => {
         const checkIfMeasuring = () => {
-          if (this.autoMeasurementState === 'idle') {
+          if (!this.gridRef.state.hasPendingMeasurements) {
             window.RESIZE_MEASUREMENT_DONE = true;
           } else {
             window.requestAnimationFrame(checkIfMeasuring);
@@ -114,10 +112,6 @@ export default class MasonryExample extends React.Component {
 
   getScrollContainerRef = ref => {
     this.scrollContainer = ref;
-  };
-
-  handleAutoMeasuringUpdate = state => {
-    this.autoMeasurementState = state;
   };
 
   handleToggleScrollContainer = e => {
@@ -330,7 +324,6 @@ export default class MasonryExample extends React.Component {
             flexible={Boolean(this.props.flexible)}
             items={this.state.items}
             measurementStore={this.props.externalCache ? store : undefined}
-            onAutoMeasuringUpdate={this.handleAutoMeasuringUpdate}
             ref={ref => {
               this.gridRef = ref;
             }}
